Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.1.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
reportlab==4.0.7
openpyxl==3.1.2
celery==5.3.4
redis==5.0.1